\name{btutils-package}
\alias{btutils-package}
\alias{btutils}
\docType{package}
\title{
Utilities for simple back testing of trading strategies.
}
\description{
A framework to test the impact of risk and money management
trading techniques (stop loss, profit target, to name a few)
on a trading strategy.
}
\details{
\tabular{ll}{
Package: \tab btutils\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-07\cr
License: \tab BSD_2_clause\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ivan Popivanov

Maintainer: Ivan Popivanov <ivan.popivanov@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
